+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Feature: Add support for variant filtering. (#140)
+
 # 2.0.1
 
 * Fix: Only upload debug symbols for non debuggable App. (#139)